removed from space single function class plotter.py into agents/plotter since it is an agent query, the function itself brute force loop over agents to find the room will be optmized to use a sorted room array
