+++ python-code/src/amr_hub_abm/agent/plotter.py
@@ -9,12 +9,14 @@ from amr_hub_abm.agent.enums import (
     ROLE_COLOUR_MAP,
     InfectionStatus,
 )
+from amr_hub_abm.exceptions import InvalidRoomError
 from amr_hub_abm.task.task import TaskAttendPatient, TaskProgress
 
 if TYPE_CHECKING:
     from matplotlib.axes import Axes
 
     from amr_hub_abm.agent.agent import Agent
+    from amr_hub_abm.space.room import Room
 
 
 def plot_agent(agent: Agent, ax: Axes, *, show_tags: bool = True) -> None:
@@ -126,3 +128,27 @@ def plot_trajectory(agent: Agent, ax: Axes, current_time: int | None = None) -> 
         color=ROLE_COLOUR_MAP[agent.agent_type],
         alpha=0.7,
     )
+
+
+def plot_agents_in_room(
+    room: Room,
+    ax: Axes,
+    agents: list[Agent],
+    *,
+    trajectory: bool = False,
+) -> None:
+    """Plot agents located inside a room."""
+    if not room.walls:
+        msg = "Cannot plot room without walls. Please provide wall definitions."
+        raise InvalidRoomError(msg)
+
+    for agent in agents:
+        if (
+            agent.location.building == room.building
+            and agent.location.floor == room.floor
+            and room.contains_point((agent.location.x, agent.location.y))
+        ):
+            plot_agent(agent, ax)
+
+            if trajectory and agent.agent_type.value == 2:
+                plot_trajectory(agent, ax)
