+++ python-code/src/amr_hub_abm/space/floor.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
 
+from amr_hub_abm.agent.plotter import plot_agents_in_room
 import numpy as np
 
 from amr_hub_abm.exceptions import InvalidRoomError
-from amr_hub_abm.space.plotter import plot_agents_in_room
 from amr_hub_abm.space.room import Room
 from amr_hub_abm.space.wall import Wall
